Feat: 게시글 삭제전 게시글 존재 여부 확인

+++ src/main/java/com/gdscdeu/springboot/simpleboard/Controllers/PostController.java
@@ -39,7 +39,14 @@ public class PostController {
     }
     
     @DeleteMapping("/{postID}")
-    void deletePost(@PathVariable long postID) {
+    void deletePost(@PathVariable long postID, HttpServletResponse response) {
+        Posts post = postService.getPost(postID);
+
+        if (post == null) {
+            response.setStatus(HttpServletResponse.SC_NOT_FOUND);
+            return;
+        }
+
         postService.deletePost(postID);
     }
